pyproject.toml
[project]
name = "tg-ip-locator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "python-nmap>=0.7.1",
    "requests>=2.32.3",
]
